time

+++ src/Dinic.java
@@ -88,7 +88,7 @@ public class Dinic extends Algoritmo {
 		  }
 
 		  public Solucion ejecutar(Entrada e) {
-			double t1 = System.currentTimeMillis();
+			long t1 = System.currentTimeMillis();
 		    int flow = 0;
 			GrafoAntiguo g = e.consultarGrafo();
 			int src = e.consultarSource();
@@ -111,7 +111,7 @@ public class Dinic extends Algoritmo {
 				crearItinerarios(sol,g,0,flow-1,flow,src,dest,0);
 			}
 		//	Guardar(path,file); 
-			double t2 = System.currentTimeMillis();
+			long t2 = System.currentTimeMillis();
 			sol.modificarTiempo(t2-t1);
 			sol.seqsol = seq; 
 			return sol;
